added prompt

+++ admtools.py
@@ -1,4 +1,6 @@
 import os
+import win32api
+import win32con
 import pathDir
 import subprocess
 from PyQt5.QtWidgets import QPushButton, QWidget
@@ -32,10 +34,16 @@ def enable_win11_Mcontext():  # Enable Windows 11 Menu Context
         exeXc(command, "Windows 11 Menu Context enabled successfully.")
 
 def disable_win11_Mcontext():  # Disable Windows 11 Menu Context to the old Windows 10 version
+    print("Waiting for user input...")
+    ret = win32api.MessageBox(0, "The Windows Explorer will be restarted...", "Warning", win32con.MB_OKCANCEL)
+        
+    if ret == win32con.IDOK:     
         command = r'reg add "HKCU\Software\Classes\CLSID\{86ca1aa0-34aa-4e8b-a509-50c905bae2a2}\InprocServer32" /f /ve'
         exeXc(command, "Windows 11 Menu Context disabled successfully.")  
         subprocess.run("taskkill /f /im explorer.exe", shell=True)
-        subprocess.run("start explorer.exe", shell=True)  
+        subprocess.run("start explorer.exe", shell=True)
+    else:
+        print("Operation cancel by user")
